[FIX] portal, purchase, sale: fix sidebar menu
Before this commit the menu built in the sidebar would not be
created properly, which could result in bugs or just one item
behing present.

This commit moves the main functions that generate the sidebar
menu in the base interaction to avoid duplicate, modify it to
handle the style that was added by the purchase sidebar and
make sure all the text content is taken into account for the
menu.

Moreover it improves the master version by not splitting some
titles. Before, "Request for Quotations P000004" was splitted
in 2 items ("Request ..." and "P000004"). It is now only 1 item.

a
